*Fixed a VRAM inspecting bug. *Improved calculation of scrollbar visibility.

+++ src/app/vram_debugger.h
@@ -21,7 +21,7 @@ public:
 	virtual bool open(class Renderer* rnd, class Theme* theme) = 0;
 	virtual bool close(void) = 0;
 
-	virtual float safeHeight(void) const = 0;
+	virtual int safeHeight(void) const = 0;
 
 	virtual int highlightCount(void) const = 0;
 	/**
